refactor: extract serve_file
